% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{create_lag}
\alias{create_lag}
\title{Create lagged variable.}
\usage{
create_lag(df, .var, .year, ...)
}
\arguments{
\item{df}{: dataframe containing year variable.}

\item{.var}{: variable to lag.}

\item{.year}{: name of time variable which we will sort by.}

\item{...}{: additional grouping variables.}
}
\value{
dataframe with lagged variable
}
\description{
Create lagged variable.
}
