% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{drop_weird_obs}
\alias{drop_weird_obs}
\title{Drop observations with weird values.}
\usage{
drop_weird_obs(df_fgy)
}
\arguments{
\item{df_fgy}{: dataframe unit of observation = (firm, good, year).
This dataframe should contain the following columns:
\itemize{
\item q: Quantity.
\item p: Price.
}}
}
\value{
a dataframe with weird observations dropped.
}
\description{
Drop observations with weird values.
}
