% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{get_market_share}
\alias{get_market_share}
\title{Compute market share for each firm-year.}
\usage{
get_market_share(df, rev_var, sector)
}
\arguments{
\item{df}{: dataframe unit of observation = (firm, year).
This dataframe should contain the following columns:
\itemize{
\item firm_id: Firm identifier.
\item year: Year.
\item rev_var: Revenue variable.
\item sector: Sector variable.
}}

\item{rev_var}{: (str) name of the revenue variable.}

\item{sector}{: (str) name of the sector variable.}
}
\value{
a vector of market shares.
}
\description{
Compute market share for each firm-year.
}
\examples{
get_market_share(d_fek, "value_added", "sni4");
}
