% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{get_markup}
\alias{get_markup}
\title{Compute markup for each firm-year.}
\usage{
get_markup(o_elas, revenue, var_cost)
}
\arguments{
\item{o_elas}{: output elasticity of the variable input.}

\item{revenue}{: total revenue, P_(i,t) * Y_(i,t).}

\item{var_cost}{: variable input total expenditure, W_t * V_(i,t).}
}
\value{
a vector of markups.
}
\description{
This follows the production approach to markups from Hall (1986, 1988).
The variable input is assumed to be competitive, hence t-subscript on W.
}
\examples{
mu <- get_markup(0.5, c(100, 200), c(70, 140));
}
