% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{get_output_elasticity}
\alias{get_output_elasticity}
\title{Compute the output elasticity of a competitive variable input}
\usage{
get_output_elasticity(df_fy)
}
\arguments{
\item{df}{: dataframe unit of observation = (firm, year).
This dataframe should contain the following columns:
\itemize{
\item firm_id: Firm identifier.
\item year: Year.
\item p: Price.
\item s: market share.
\item w: price of the variable input. TODO how to handle this? Just subtract a mean?
\item v: quantity of the variable input.
\item v_lag: quantity of the variable input in the previous period.
\item y: output in physical units.
\item NB: These variables all need to be log deviations from their time-mean.
}}
}
\value{
an output elasticity.
}
\description{
This follows the procedure in De Ridder, Grassi, Morzenti (2022) sec. 3.
TODO: adapt to case where price data is not present.
TODO: group by so different groups have different output elasticities.
}
