% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{log_dev_from_mean}
\alias{log_dev_from_mean}
\title{Compute log deviation from mean of a variable.}
\usage{
log_dev_from_mean(var)
}
\arguments{
\item{var}{: (str) variable to be demeaned.}
}
\value{
a variable : log deviation from mean of var.
}
\description{
This follows the procedure in De Ridder, Grassi, Morzenti (2022)
sec. 3.1. footnote 10.
}
