% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{standardize_price}
\alias{standardize_price}
\title{Standardize prices following De Ridder, Grassi, Morzenti (2022).}
\usage{
standardize_price(df_fgy, good)
}
\arguments{
\item{df_fgy}{: Dataframe with firm-year-product observations.
This dataframe should contain the following columns:
\itemize{
\item firm_id: Firm identifier.
\item year: Year.
\item revenue: Revenue.
\item q: Quantity.
\item p: Price.
}}

\item{good}{: Standardize prices at the level defined by good.}
}
\value{
: df_fy, Dataframe with standardized prices id by firm and year.
}
\description{
The Hitchhiker’s Guide to Markup Estimation, section 4 "Data".
Definition of a good/product: Combination of a code and unit of account.
}
\examples{
standardize_price(d_ivp, "good_id")

}
