% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_functions.R
\name{winsorize_vars}
\alias{winsorize_vars}
\title{Winsorize variables.}
\usage{
winsorize_vars(.df, .var_list, ...)
}
\arguments{
\item{.df}{: dataframe.}

\item{.var_list}{: list of variables to winsorize.}

\item{...}{: grouping variables. dRGM (2022) group by 2 digit industry.}
}
\value{
a dataframe with winsorized variables.
}
\description{
Winsorize variables.
}
